+++ exporters/otlp/otlpmetric/internal/exporter.go
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package internal provides common utilities for all otlpmetric exporters.
+//
+// Deprecated: package internal exists for historical compatibility, it should
+// not be used.
 package internal // import "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal"
 
 import (
@@ -19,19 +23,13 @@ import (
 	"fmt"
 	"sync"
 
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/transform"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/transform" // nolint: staticcheck  // Atomic deprecation.
 	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	mpb "go.opentelemetry.io/proto/otlp/metrics/v1"
 )
 
 // Exporter exports metrics data as OTLP.
-//
-// Deprecated: Exporter exists for historical compatibility, it should not be
-// used. Do not remove Exporter unless the whole
-// "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal" module is
-// removed.
 type Exporter struct {
 	// Ensure synchronous access to the client across all functionality.
 	clientMu sync.Mutex
@@ -48,7 +46,7 @@ func (e *Exporter) Temporality(k metric.InstrumentKind) metricdata.Temporality {
 }
 
 // Aggregation returns the Aggregation to use for an instrument kind.
-func (e *Exporter) Aggregation(k metric.InstrumentKind) aggregation.Aggregation {
+func (e *Exporter) Aggregation(k metric.InstrumentKind) metric.Aggregation {
 	e.clientMu.Lock()
 	defer e.clientMu.Unlock()
 	return e.client.Aggregation(k)
@@ -101,11 +99,6 @@ func (e *Exporter) Shutdown(ctx context.Context) error {
 // New return an Exporter that uses client to transmits the OTLP data it
 // produces. The client is assumed to be fully started and able to communicate
 // with its OTLP receiving endpoint.
-//
-// Deprecated: New exists for historical compatibility, it should not be used.
-// Do not remove New unless the whole
-// "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal" module is
-// removed.
 func New(client Client) *Exporter {
 	return &Exporter{client: client}
 }
@@ -126,7 +119,7 @@ func (c shutdownClient) Temporality(k metric.InstrumentKind) metricdata.Temporal
 	return c.temporalitySelector(k)
 }
 
-func (c shutdownClient) Aggregation(k metric.InstrumentKind) aggregation.Aggregation {
+func (c shutdownClient) Aggregation(k metric.InstrumentKind) metric.Aggregation {
 	return c.aggregationSelector(k)
 }