+++ internal/tools/tools.go
@@ -29,5 +29,6 @@ import (
 	_ "go.opentelemetry.io/build-tools/gotmpl"
 	_ "go.opentelemetry.io/build-tools/multimod"
 	_ "go.opentelemetry.io/build-tools/semconvgen"
+	_ "golang.org/x/exp/cmd/gorelease"
 	_ "golang.org/x/tools/cmd/stringer"
 )