+++ sdk/metric/internal/aggregate/histogram_test.go
@@ -23,17 +23,13 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
 )
 
 var (
 	bounds   = []float64{1, 5}
-	histConf = aggregation.ExplicitBucketHistogram{
-		Boundaries: bounds,
-		NoMinMax:   false,
-	}
+	noMinMax = false
 )
 
 func TestHistogram(t *testing.T) {
@@ -59,7 +55,7 @@ func testDeltaHist[N int64 | float64](c conf[N]) func(t *testing.T) {
 	in, out := Builder[N]{
 		Temporality: metricdata.DeltaTemporality,
 		Filter:      attrFltr,
-	}.ExplicitBucketHistogram(histConf, c.noSum)
+	}.ExplicitBucketHistogram(bounds, noMinMax, c.noSum)
 	ctx := context.Background()
 	return test[N](in, out, []teststep[N]{
 		{
@@ -125,7 +121,7 @@ func testCumulativeHist[N int64 | float64](c conf[N]) func(t *testing.T) {
 	in, out := Builder[N]{
 		Temporality: metricdata.CumulativeTemporality,
 		Filter:      attrFltr,
-	}.ExplicitBucketHistogram(histConf, c.noSum)
+	}.ExplicitBucketHistogram(bounds, noMinMax, c.noSum)
 	ctx := context.Background()
 	return test[N](in, out, []teststep[N]{
 		{
@@ -277,7 +273,7 @@ func TestHistogramImmutableBounds(t *testing.T) {
 	cpB := make([]float64, len(b))
 	copy(cpB, b)
 
-	h := newHistogram[int64](aggregation.ExplicitBucketHistogram{Boundaries: b}, false)
+	h := newHistogram[int64](b, false, false)
 	require.Equal(t, cpB, h.bounds)
 
 	b[0] = 10
@@ -293,7 +289,7 @@ func TestHistogramImmutableBounds(t *testing.T) {
 }
 
 func TestCumulativeHistogramImutableCounts(t *testing.T) {
-	h := newHistogram[int64](histConf, false)
+	h := newHistogram[int64](bounds, noMinMax, false)
 	h.measure(context.Background(), 5, alice)
 
 	var data metricdata.Aggregation = metricdata.Histogram[int64]{}
@@ -311,7 +307,7 @@ func TestCumulativeHistogramImutableCounts(t *testing.T) {
 func TestDeltaHistogramReset(t *testing.T) {
 	t.Cleanup(mockTime(now))
 
-	h := newHistogram[int64](histConf, false)
+	h := newHistogram[int64](bounds, noMinMax, false)
 
 	var data metricdata.Aggregation = metricdata.Histogram[int64]{}
 	require.Equal(t, 0, h.delta(&data))
@@ -340,21 +336,21 @@ func BenchmarkHistogram(b *testing.B) {
 	b.Run("Int64/Cumulative", benchmarkAggregate(func() (Measure[int64], ComputeAggregation) {
 		return Builder[int64]{
 			Temporality: metricdata.CumulativeTemporality,
-		}.ExplicitBucketHistogram(histConf, false)
+		}.ExplicitBucketHistogram(bounds, noMinMax, false)
 	}))
 	b.Run("Int64/Delta", benchmarkAggregate(func() (Measure[int64], ComputeAggregation) {
 		return Builder[int64]{
 			Temporality: metricdata.DeltaTemporality,
-		}.ExplicitBucketHistogram(histConf, false)
+		}.ExplicitBucketHistogram(bounds, noMinMax, false)
 	}))
 	b.Run("Float64/Cumulative", benchmarkAggregate(func() (Measure[float64], ComputeAggregation) {
 		return Builder[float64]{
 			Temporality: metricdata.CumulativeTemporality,
-		}.ExplicitBucketHistogram(histConf, false)
+		}.ExplicitBucketHistogram(bounds, noMinMax, false)
 	}))
 	b.Run("Float64/Delta", benchmarkAggregate(func() (Measure[float64], ComputeAggregation) {
 		return Builder[float64]{
 			Temporality: metricdata.DeltaTemporality,
-		}.ExplicitBucketHistogram(histConf, false)
+		}.ExplicitBucketHistogram(bounds, noMinMax, false)
 	}))
 }